backend/requirements.txt
Flask
pymongo
python-dotenv
openai
